finish runtime
31 percent, woo

+++ include/PowerPC_EABI_Support/Runtime/runtime.h
@@ -7,7 +7,7 @@
 extern "C" {
 #endif // ifdef __cplusplus
 
-unsigned long __cvt_fp2unsigned(f64);
+u32 __cvt_fp2unsigned(f64);
 // TODO: The rest
 
 void* __copy(char*, char*, size_t);
